move data fetching up to page level
