+++ src/pages/index.tsx
@@ -1,11 +1,52 @@
+import { useState, useEffect } from 'react'
 import Page from '@/components/side-effects/Page'
 import withAuth from '@/components/side-effects/withAuth'
 import ProjectsTable from '@/components/side-effects/ProjectsTable'
+import { defaultFilters, defaultSort } from '@/components/page/overview/types'
+import {
+  Project,
+  ProjectFilter,
+  ProjectOrderBy,
+  useTrendingProjectsQuery
+} from '@/graphql/generated/gql'
 
-const Home = () => (
-  <Page>
-    <ProjectsTable />
-  </Page>
-)
+const TrendingProjects = () => {
+  const [data, setData] = useState<Project[]>([])
+  const [filters, setFilters] = useState<ProjectFilter>(defaultFilters)
+  const [sorting, setSorting] = useState<ProjectOrderBy | null>(defaultSort)
 
-export default withAuth(Home)
+  const updateFilters = (filter: ProjectFilter) => {
+    setFilters(filter)
+  }
+
+  // Fetch data from Supabase using generated Urql hook
+  const [{ data: urqlData, fetching, error }] = useTrendingProjectsQuery({
+    variables: {
+      orderBy: sorting || defaultSort,
+      filter: filters || defaultFilters
+    }
+  })
+
+  // Only update table data when urql data changes
+  useEffect(() => {
+    if (urqlData) {
+      setData(urqlData?.projectCollection?.edges?.map((edge) => edge.node) as Project[])
+    }
+  }, [urqlData])
+
+  return (
+    <Page>
+      <ProjectsTable
+        data={data}
+        filters={filters}
+        sorting={sorting}
+        fetching={fetching}
+        error={error}
+        setSorting={setSorting}
+        updateFilters={updateFilters}
+      />
+    </Page>
+  )
+}
+
+export default withAuth(TrendingProjects)
